Update CI, add docs generation
Also rearrange tests.
