+++ README.md
@@ -1,8 +1,16 @@
 # MeasureBase
 
-[![Stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://JuliaMath.github.io/MeasureTheory.jl/stable)
-[![Dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://JuliaMath.github.io/MeasureTheory.jl/dev)
-[![Build Status](https://github.com/JuliaMath/MeasureBase.jl/workflows/CI/badge.svg)](https://github.com/JuliaMath/MeasureBase.jl/actions)
-[![Coverage](https://codecov.io/gh/JuliaMath/MeasureBase.jl/branch/master/graph/badge.svg)](https://codecov.io/gh/JuliaMath/MeasureBase.jl)
+[![Documentation for stable version](https://img.shields.io/badge/docs-stable-blue.svg)](https://JuliaMath.github.io/MeasureBase.jl/stable)
+[![Documentation for development version](https://img.shields.io/badge/docs-dev-blue.svg)](https://JuliaMath.github.io/MeasureBase.jl/dev)
+[![License](http://img.shields.io/badge/license-MIT-brightgreen.svg?style=flat)](LICENSE.md)
+[![Build Status](https://github.com/JuliaMath/MeasureBase.jl/workflows/CI/badge.svg?branch=main)](https://github.com/JuliaMath/MeasureBase.jl/actions?query=workflow%3ACI)
+[![Codecov](https://codecov.io/gh/JuliaMath/MeasureBase.jl/branch/main/graph/badge.svg)](https://codecov.io/gh/JuliaMath/MeasureBase.jl)
+[![Aqua QA](https://raw.githubusercontent.com/JuliaTesting/Aqua.jl/master/badge.svg)](https://github.com/JuliaTesting/Aqua.jl)
 
-This (relatively) light-weight package contains core functionality of [MeasureTheory](https://github.com/JuliaMath/MeasureTheory.jl). Many packages using or defining measures do not need the full capabilities of MeasureTheory.jl itself, and can depend on this instead.
+
+## Documentation
+
+* [Documentation for stable version](https://JuliaMath.github.io/MeasureBase.jl/stable)
+* [Documentation for development version](https://JuliaMath.github.io/MeasureBase.jl/dev)
+
+This package contains core definitions for (probability and non-probability) measures.
