+++ docs/src/LICENSE.md
@@ -1,0 +1,6 @@
+# LICENSE
+
+```@eval
+using Markdown
+Markdown.parse_file(joinpath(@__DIR__, "..", "..", "LICENSE.md"))
+```
