+++ docs/src/api.md
@@ -1,0 +1,26 @@
+# API
+
+## Modules
+
+```@index
+Order = [:module]
+```
+
+## Types and constants
+
+```@index
+Order = [:type, :constant]
+```
+
+## Functions and macros
+
+```@index
+Order = [:macro, :function]
+```
+
+# Documentation
+
+```@autodocs
+Modules = [MeasureBase]
+Order = [:module, :type, :constant, :macro, :function]
+```
