+++ docs/src/index.md
@@ -1,0 +1,3 @@
+# MeasureBase.jl
+
+This package contains core definitions for (probability and non-probability) measures.
